Add google maps to mappa.html

+++ php/delete_report.php
@@ -1,11 +1,16 @@
 <?php
   include "db_connection.php";
   session_start();
-  $reportID = $_GET['id'];
-  $sql = "DELETE FROM t_segnalazioni WHERE ID='$reportID'";
-  $deleteQuery = mysqli_query($db_conn, $sql);
-  if ($deleteQuery == null){
-    die("error");
+  if ($_SESSION['isLogged']){
+    $reportID = $_GET['id'];
+    $sql = "DELETE FROM t_segnalazioni WHERE ID='$reportID'";
+    $deleteQuery = mysqli_query($db_conn, $sql);
+    if ($deleteQuery == null){
+      die("error");
+    }
+    header("location:../dashboard.php");
+  }else{
+    header("location:../login.php");
   }
-  header("location:../dashboard.php");
+
 ?>
